fix abstraction over auth and create first test
