+++ tests/fixtures/server/stats-new.json
