+++ tests/fixtures/server/stats.json
